[x64] minor test-only updates for better type safety

+++ tests/array_test.py
@@ -293,7 +293,7 @@ class JaxArrayTest(jtu.JaxTestCase):
     a, input_data = create_array(
         input_shape, sharding.NamedSharding(global_mesh, P('x', 'y')))
     out = jnp.zeros_like(a)
-    expected = jnp.zeros(input_data.shape, dtype=int)
+    expected = jnp.zeros(input_data.shape, dtype=a.dtype)
     self.assertArraysEqual(out, expected)
     self.assertLen(out.addressable_shards, 8)
     for i in out.addressable_shards:
